+++ src/extension-firefox/toggle.js
@@ -1,0 +1,24 @@
+/* toggle.js */
+
+(function () {
+
+  const debug = false;
+
+  debug && console.log(`[toggle.js][onclick]`);
+
+  let skipToContentElem = document.querySelector('skip-to-content');
+
+  debug && console.log(`[toggle.js]: ${skipToContentElem} ${skipToContentElem.checked}`);
+
+  if (skipToContentElem) {
+    const focus = skipToContentElem.getAttribute('focus');
+    if (focus === 'menu') {
+      skipToContentElem.setAttribute('setfocus', 'none');
+    }
+    else {
+      skipToContentElem.setAttribute('setfocus', 'menu');
+    }
+  }
+
+})();
+